MediaGenieBot/requirements.txt
python-telegram-bot==20.3
aiohttp
requests
yt_dlp
